See your SQL queries on the SQL tab when using your existing NHibernate assemblies in LINQPad
Ordinarily, if you manage to get your NHibernate assemblies to run in
LINQPad, the generated SQL is difficult to obtain, doesn't appear on the
SQL tab and can't easily be run without being manually edited.

This fixes all of that, it has a demo with a SQLite database and it
works in the trial version of LINQPad.